requirements.txt
pyTelegramBotAPI
cryptography
pyzipper
python-dotenv
qrcode[pil]
pillow
schedule
matplotlib
psutil
pytest
